Working on implementing shooting system; type issues at the moment

+++ src/entities/Enemy.cpp
@@ -118,7 +118,7 @@ int Enemy::dealDamage() const
     return attackDamage;
 }
 
-void Enemy::takeDamage(int damage)
+void Enemy::takeDamage(float damage)
 {
     (void)damage;
